+++ CHANGELOG.md
@@ -2,7 +2,24 @@
 
 (or: why are you updating this all the time?)
 
-### Newest version: 1.2.9
+### Newest version: 1.3.3
+
+* The show / hide methods now behave nicely with CoordinatorLayout.
+* Added alpha animation for the tab titles when using the shifting tabs.
+
+### 1.3.2
+
+* Now it's possible to use fixed mode (show titles on inactive tabs) even when there's more than three tabs.
+
+### 1.3.1
+
+* Fixed a critical bug in OnLongClickListener behavior (why didn't I see that before?) when using badges.
+
+### 1.3.0
+
+* Fixed a critical bug in OnClickListener behavior when using badges.
+
+### 1.2.9
 
 * Fixed the issue when using badges and the tab resize animation is enabled. Now the badges automatically adjust their position when the tab's size (or position) updates.